% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationRuleMining.R
\docType{package}
\name{AssociationRuleMining-package}
\alias{AssociationRuleMining}
\alias{AssociationRuleMining-package}
\title{AssociationRuleMining: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Solomon Ioannou \email{s.ioannou@erasmusmc.nl} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
