% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndromedaHelperFunctions.R
\name{convertAndromedaToArrow}
\alias{convertAndromedaToArrow}
\title{convertAndromedaToArrow}
\usage{
convertAndromedaToArrow(andromeda, path)
}
\arguments{
\item{andromeda}{andromeda table such as plpData$covariateData$covariates}

\item{path}{where to save arrow dataset}
}
\description{
converts an andromeda table to arrow dataset and writes it to the specified path
}
