% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndromedaHelperFunctions.R
\name{convertArrowToCyclopsData}
\alias{convertArrowToCyclopsData}
\title{convertArrowToCyclopsData}
\usage{
convertArrowToCyclopsData(
  outcomes,
  covariates,
  modelType = "lr",
  addIntercept = TRUE,
  checkSorting = NULL,
  checkRowIds = TRUE,
  normalize = NULL,
  quiet = FALSE,
  floatingPoint = 64
)
}
\description{
convertArrowToCyclopsData
}
\section{Functions}{
\itemize{
\item \code{convertArrowToCyclopsData}: Convert data from two \code{Dataset} tables
}}

