% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AndromedaHelperFunctions.R
\name{convertCovariateDataToArrow}
\alias{convertCovariateDataToArrow}
\title{convertCovariateDataToArrow}
\usage{
convertCovariateDataToArrow(covariateData)
}
\arguments{
\item{covariateData}{A covariateData object from featureExtraction}
}
\description{
Converts a covariateData object from using a rsqlite backend
to arrow backend
}
