% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationRulesFeatureEngineering.R
\name{createAssociationRulesSettings}
\alias{createAssociationRulesSettings}
\title{Settings for association rule mining}
\usage{
createAssociationRulesSettings(
  parameters,
  control,
  removeLengthOnePatterns = FALSE,
  transactionsPlpData,
  savePatterns = FALSE
)
}
\arguments{
\item{parameters}{An object of class APparameter for apriori (or ECparameter for eclat) or a named list. Available slots
\itemize{
\item APparameter: confidence, minval, smax, arem, aval, originalSupport, maxtime, support, minlen, maxlen, target, ext.
\item ECparameter: tidLists, support, minlen, maxlen, target, ext
}}

\item{control}{An object of class APcontrol for apriori (or ECcontrol for ecalt) or named list. Available slots
\itemize{
\item APcontrol: filter, tree, heap, memopt, load, sort, verbose
\item ECcontrol: sparse, sort, verbose
}}

\item{removeLengthOnePatterns}{Whether to remove patterns (rules) with 1 item.}

\item{savePatterns}{Whether to save mined patterns (rules) or not. Default to FALSE.}

\item{trasnactionsPlpData}{A plpData object to convert to transactions for mining}
}
\description{
Settings for association rule mining
}
