% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatabaseSummaryForFrequentPatterns.R
\name{getDatabaseSummaryForFrequentPatterns}
\alias{getDatabaseSummaryForFrequentPatterns}
\title{Database summary with parameter description for running Frequent Pattern Mining algorithms}
\usage{
getDatabaseSummaryForFrequentPatterns(data = data)
}
\arguments{
\item{data}{A tibble or a data.frame object as returned by the \code{getInputFileForFrequentPatterns()}.}
}
\value{
A list with tibbles.
}
\description{
\code{getDatabaseSummary} returns descriptive statistics for the input dataset
}
\details{
This function gives descriptive statistics for various parameters: number of sequences, length of sequences and size of events in sequences.
}
