% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdDataFrame.R
\name{getIdDataFrame}
\alias{getIdDataFrame}
\title{Get a dataframe indicating which patterns are observed in each patient}
\usage{
getIdDataFrame(inputFile, objectWithIds)
}
\arguments{
\item{inputFile}{The .txt file returned from \code{runFrequentPatterns()}.}

\item{objectWithIds}{The tibble returned from \code{getInputFileFromFrequentPatterns()}.}
}
\description{
\code{getIdDataFrame} returns a data frame with patients as rows and frequent patterns as columns.
}
