% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInputFileForFrequentPatterns.R
\name{getInputFileForFrequentPatterns}
\alias{getInputFileForFrequentPatterns}
\title{Prepare input data for JAVA algorithms}
\usage{
getInputFileForFrequentPatterns(covariateDataObject, fileToSave)
}
\arguments{
\item{covariateDataObject}{A covariateData object with temporal information as returned by \code{FeatureExtraction}.}

\item{fileToSave}{Location and name of file to save the input data. Should be a .txt file.}
}
\value{
A tibble and a .txt file at the location specified.
}
\description{
\code{getInputFileForFrequentPatterns} returns a tibble with input data in long format and a .txt file at the directory specified.
}
\details{
The function takes a covariateData object as returned by the \code{FeatureExtraction} package and a location or name of file (should be .txt file) to save the input data.
}
