% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{prepareSankey}
\alias{prepareSankey}
\title{Prepares data to visualise frequent patterns with a sankey diagram with networkD3.}
\usage{
prepareSankey(data = data, plot = c("Sankey"), inputFile, objectWithIds)
}
\arguments{
\item{data}{A tibble as returned from \code{runFrequentPatterns()}. It makes more sense to use one of the algorithms that result
in the closed or maximal sets.}

\item{inputFile}{The .txt file defined as the \code{outputFile} from \code{runFrequentPatterns()}.}

\item{objectWithIds}{The tibble returned from \code{getInputFileFromFrequentPatterns()}.}
}
\description{
\code{prepareSankey} returns a list with nodes and links.
}
