% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessing.R
\name{reduceClosedPatterns}
\alias{reduceClosedPatterns}
\title{Reducing length of Frequent Patterns}
\usage{
reduceClosedPatterns(inputFile)
}
\arguments{
\item{data}{A tibble as returned from \code{runFrequentPatterns()}.}
}
\description{
\code{reduceClosedPatterns} Removes repetitive concepts from Frequent Patterns.
}
\details{
It takes as input the tibble resulted from \code{runFrequentPatterns()} and removes repetitive concepts from the extracted patterns
}
