% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runFrequentPatterns.R
\name{runFrequentPatterns}
\alias{runFrequentPatterns}
\title{Run Frequent Patterns algorithms}
\usage{
runFrequentPatterns(
  algorithm,
  inputFile,
  outputFile,
  minsup,
  minconf,
  minLength = 1,
  maxLength = Inf,
  maxGap = Inf,
  maxAntecedentLength = Inf,
  maxConsequentLength = 1,
  showID = FALSE
)
}
\arguments{
\item{algorithm}{One between "SPAM", "SPADE", "prefixSpan", "CM-SPADE", "CM-SPAM", "Clasp", "CM-Clasp", "BIDE+", "VMSP", "VGEN", "RuleGrowth", "ERMiner", "TNS", "TopSeqRules".}

\item{inputFile}{Name of the .txt file to be processed.}

\item{outputFile}{Name of .txt file to extract results.}

\item{minsup}{Value of minimum support between 0-1.}

\item{minconf}{Value of minimum confidence. Only some algorithms are making use of this (TO be specified)}
}
\description{
\code{runFrequentPatterns} returns the extracted frequent patterns.
}
\details{
The function takes a .txt file as input and returns the results in a .txt file as well as a tibble.
}
